Move NodeModuleGenerator.test.ts and ReadmeGenerator.test.ts and support files to test dir
