+++ src/app/TypeScriptPackageGenerator.test.ts
@@ -1,7 +1,7 @@
 import YeomanHelpers, { RunResult } from "yeoman-test";
 
+import { NODE_MODULE_GENERATOR_TEST_OPTIONS } from "../../test/fixtures";
 import { RunResultUtils } from "../../test/utils";
-import { NODE_MODULE_GENERATOR_TEST_OPTIONS } from "../node-module/__tests__/__fixtures__/index";
 
 describe("TypeScriptPackageGenerator", () => {
   let result: RunResult;
