+++ src/node-module/index.ts
@@ -1,2 +1,3 @@
 export { NodeModuleGenerator as default } from "./NodeModuleGenerator";
 export * from "./NodeModuleGenerator";
+export * from "./readme";
