--- src/node-module/readme/__tests__/__fixtures__/index.ts
@@ -1,1 +1,0 @@
-export * from "./options";
