+++ test/fixtures/generatorOptions/index.ts
@@ -1,0 +1,2 @@
+export * from "./nodeModuleGeneratorTest.options";
+export * from "./readmeGeneratorTest.options";
