+++ test/unit/node-module/NodeModuleGenerator.test.ts
@@ -1,7 +1,7 @@
 import path from "node:path";
 import YeomanHelpers, { RunResult } from "yeoman-test";
 
-import { NODE_MODULE_GENERATOR_TEST_OPTIONS } from "./__tests__/__fixtures__/";
+import { NODE_MODULE_GENERATOR_TEST_OPTIONS } from "../../fixtures";
 
 describe("NodeModuleGenerator", () => {
   let result: RunResult;
@@ -223,5 +223,7 @@ describe("NodeModuleGenerator", () => {
 });
 
 const runWithAnswers = (answers: Record<string, any>): Promise<RunResult> => {
-  return YeomanHelpers.create(path.join(__dirname)).withPrompts(answers).run();
+  return YeomanHelpers.create(path.join(__dirname, "../../../src/node-module"))
+    .withPrompts(answers)
+    .run();
 };
